Fix select call's timeout logic for linux.
"On  Linux, select() modifies timeout to reflect the amount of time not slept;
most other implementations do not do this.  (POSIX.1-2001 permits either
behavior.)  This causes problems both when Linux code which reads timeout is
ported to other operating systems, and when code is ported to Linux that reuses
a struct timeval for multiple select()s in a loop without reinitializing it.
Consider timeout to be undefined after select() returns."